test/unit: cover all the combinations
